day7_echo_bot_v2/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.27.0
python-dotenv==1.0.1
jinja2==3.1.6
pydantic==2.11.7
requests==2.32.4 
assemblyai
